+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 Changelog
 =========
 
+## 6.20.0 (2022-07-13)
+
+### Enhancements
+
+* Feature flags can now be accessed in the onSend callbacks.
+  [#1432](https://github.com/bugsnag/bugsnag-cocoa/pull/1432)
+
+* Capture `userInfo` from all NSExceptions and include the `error` metadata tab for handled exceptions.
+  [#1428](https://github.com/bugsnag/bugsnag-cocoa/pull/1428)
+
+* Feature flags are now kept in order of insertion or modification rather than in alphabetical order.
+  [#1429](https://github.com/bugsnag/bugsnag-cocoa/pull/1429)
+
+* Send usage telemetry to Bugsnag for product improvement purposes. Can be disabled using `configuration.telemetry`.
+  [#1422](https://github.com/bugsnag/bugsnag-cocoa/pull/1422)
+
+### Bug fixes
+
+* Prevent reporting of OOMs on simulators. 
+  [#1421](https://github.com/bugsnag/bugsnag-cocoa/pull/1421)
+
+* Fix a rare crash in `BugsnagBreadcrumbsWriteCrashReport()`.
+  [#1430](https://github.com/bugsnag/bugsnag-cocoa/pull/1430)
+
+* Fix intermittent empty thread stacktraces.
+  [#1425](https://github.com/bugsnag/bugsnag-cocoa/pull/1425)
+
 ## 6.19.0 (2022-06-29)
 
 ### Enhancements